consolidated cancel subscription tests
